Payu compatible with Ruby 2.4
Ruby 2.4 comes bundled with bigdecimal 1.3.2.
This version of the gem introduces breaking behavior
on the BigDecimal.new method. If we pass invalid
values now, it will raise an exception instead
of converting it to a 0 value

When testing BigDecimal 1.3.2 in Shopify we noticed that the `discount`
parameter is not always sent by `payu_in_paisa`. This created a
situation: `BigDecimal.new(nil.to_s)` which raises an error in the new
version of the gem

This commit preserves the behavior of the code before the gem bump